requirements.txt
pandas
websocket-client
google-auth
pandas_gbq
google-cloud-bigquery
mysql-connector-python
google-cloud-logging
silly